==> 09_table1_publication.qmd <==
---
title: "Table1 (for publication)"
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE,
                      eval = TRUE,
                      warnings = FALSE,
                      message = FALSE,
                      comment = "#>",
                      cache = TRUE,
                      #results = "hide",
                      digits = 4,
                      error = FALSE)

## clean the R environment
# graphics.off()
# rm(list = ls())
# freshr::freshr()

## load packages
library(here, quietly = TRUE)
library(tidyverse, quietly = TRUE)
library(gtsummary, quietly = TRUE)
library(flextable, quietly = TRUE)
library(broom.mixed)
library(nlme)
library(JMbayes2)
devtools::load_all()

## check the directory for the file
# here::dr_here()
here::set_here()

## the figure or results should be saved 
# paste0("foldername/Sfilename_workingresult_", 
#      Sys.Date(), ".filetype")
```

\newpage

## To-do-list

-   [x] run table 1 for the total sample size as 1 group

    -   [x] Keeping table1 for one group (n = 600)

    -   [x] Baseline levels of AMH (untransformed)

    -   [x] also include the min and max of AMH

    -   [x] How many repeated measures per woman

    -   [x] fmp_age and fmp_time sumamry

-   [x] all the survival parameters summary

-   [x] all the longitudinal parameters summary

-   [x] joint model summary

    -   [x] with rate

    -   [x] without rate

-   [x] `r praise::praise()`

\newpage

## Table1: descriptive summaries

Here is the Table1 for publication.

```{r "loading_data"}
load(here("data", "swan_amh1.rda"))
load(here("data", "swan_amh0.rda"))

sum_amh0 <- swan_amh0 %>%
  group_by(id) %>%
  summarize(visit_n = n(),
            amh_min = min(amh),
            amh_max = max(amh)) %>%
  ungroup() %>%
  mutate(id = as.character(id),
         visit_n = as.numeric(visit_n))

swan_amh2 <- swan_amh1 %>%
  full_join(sum_amh0, by = "id") %>%
  mutate(id = as.character(id),
         visit_n = as.numeric(visit_n))


```

```{r}
#| label: "create_table1"

table1 <- swan_amh2 %>%
  ungroup() %>%
  mutate(children = as.numeric(children),
         parity = case_when(parity == 0 ~ "None",
                            parity == 1 ~ "One",
                            parity == 2 ~ "Two",
                            parity == 3 ~ "Three",
                            parity == 4 ~ "More than three"),
         parity = factor(parity, levels = c("None", "One",
                                               "Two", "Three",
                                               "More than three"))) %>%
  dplyr::select(age, fmp_age, etime, 
                ethnic, married,
                alcohol, smoke, bc_pills,
                bmi, health, phys_act, amh,
                visit_n,  # visit,
                parity, children, 
                employed, edu, site)  %>%
  tbl_summary(# by = site,
              missing = "ifany",
              label = list(c("age") ~ "Age (years)",
                           c("fmp_age") ~ "Age of FMP (years)",
                           c("etime") ~ "Time of FMP(years)",
                           c("ethnic") ~ "Ethnicity / Race",
                           c("married") ~ "Marital Status",
                           c("alcohol") ~ "Alcohol Use",
                           c("smoke") ~ "Smoking History",
                           c("bc_pills") ~ "Ever use of BC Pills",
                           c("bmi") ~ "BMI",
                           c("health") ~ "Overall Health",
                           c("phys_act") ~ "Physical Activity Score",
                           c("amh") ~ "Baseline AMH",
                           c("parity") ~ "Parity",
                           c("children") ~ "Children",
                           c("employed") ~ "Employment Status",
                           c("edu") ~ "Educational Attainment",
                           c("visit_n") ~ "Number of Visits",
                           c("site") ~ "Site"),
              type = list(all_continuous() ~ "continuous2",
                          c("children", "visit_n") ~ "continuous2"),
              statistic = list(all_continuous() ~ c("{mean} ({sd})",
                                                    "{median} [{min}, {max}]"),
                               all_categorical() ~ "{n} / {N} ({p}%)")) %>%
  modify_header(label = "**Characteristics**") %>%
  # update the column header
  bold_labels() %>%
  # add_overall() %>%
  italicize_labels()
```

```{r}
#| label: "styled_table1"
#| tbl-cap: "Descriptive Statistics for SWAN AMH Data"
#| warning: false

set_flextable_defaults(fonts_ignore=TRUE)
# set_flextable_defaults(font.size = 8, 
#                        # text.align = "left",
#                        # theme_fun = theme_zebra,
#                        padding = 2)
table1 %>% 
  as_flex_table() %>%
  fontsize(size = 8)
  
```

Here is the LaTeX code for the table1.

```{r "table1_latex"}
#| label: "table_latex"
#| eval: false
library(xtable)
xtable(as.data.frame(table1), type = "latex",
      file = paste0("figure/swan_amh_09_publish_table1_", Sys.Date(), ".tex"))
```

```         
% latex table generated in R 4.2.2 by xtable 1.8-4 package
% Wed Apr 10 20:57:15 2024
\begin{table}[ht]
\centering
\begin{tabular}{rll}
  \hline
 & **Characteristics** & **N = 600** \\ 
  \hline
  1 & \_\_\_Age (years)\_\_\_ &  \\ 
  2 & Mean (SD) & 46.42 (2.41) \\ 
  3 & Median [Range] & 46.00 [42.00, 53.00] \\ 
  4 & \_\_\_Age of FMP (years)\_\_\_ &  \\ 
  5 & Mean (SD) & 52.98 (2.51) \\ 
  6 & Median [Range] & 53.05 [45.59, 62.10] \\ 
  7 & \_\_\_Time of FMP(years)\_\_\_ &  \\ 
  8 & Mean (SD) & 10.98 (2.51) \\ 
  9 & Median [Range] & 11.05 [3.59, 20.10] \\ 
  10 & \_\_\_Ethnicity / Race\_\_\_ &  \\ 
  11 & CAUCA & 265 / 600 (44\%) \\ 
  12 & BLACK & 170 / 600 (28\%) \\ 
  13 & CHINE & 62 / 600 (10\%) \\ 
  14 & HISPA & 18 / 600 (3.0\%) \\ 
  15 & JAPAN & 85 / 600 (14\%) \\ 
  16 & \_\_\_Marital Status\_\_\_ &  \\ 
  17 & Not Married/Partnered & 191 / 600 (32\%) \\ 
  18 & Married/Partnered & 409 / 600 (68\%) \\ 
  19 & \_\_\_Alcohol Use\_\_\_ &  \\ 
  20 & None & 308 / 600 (51\%) \\ 
  21 & $<$1/wk & 71 / 600 (12\%) \\ 
  22 & 1-7/wk & 161 / 600 (27\%) \\ 
  23 & $>$7/wk & 60 / 600 (10\%) \\ 
  24 & \_\_\_Smoking History\_\_\_ &  \\ 
  25 & Never Smoked & 373 / 600 (62\%) \\ 
  26 & Past Only & 149 / 600 (25\%) \\ 
  27 & Current Smoker & 78 / 600 (13\%) \\ 
  28 & \_\_\_Ever use of BC Pills\_\_\_ &  \\ 
  29 & Never & 155 / 600 (26\%) \\ 
  30 & Ever use BC pills & 445 / 600 (74\%) \\ 
  31 & \_\_\_BMI\_\_\_ &  \\ 
  32 & Mean (SD) & 28 (7) \\ 
  33 & Median [Range] & 26 [16, 56] \\ 
  34 & \_\_\_Overall Health\_\_\_ &  \\ 
  35 & Excellent & 147 / 600 (25\%) \\ 
  36 & Very Good & 233 / 600 (39\%) \\ 
  37 & Good & 151 / 600 (25\%) \\ 
  38 & Fair/Poor & 69 / 600 (12\%) \\ 
  39 & \_\_\_Physical Activity Score\_\_\_ &  \\ 
  40 & Mean (SD) & 7.75 (1.77) \\ 
  41 & Median [Range] & 7.60 [3.40, 12.80] \\ 
  42 & \_\_\_Baseline AMH\_\_\_ &  \\ 
  43 & Mean (SD) & 540 (702) \\ 
  44 & Median [Range] & 273 [1, 4,652] \\ 
  45 & \_\_\_Number of Visits\_\_\_ &  \\ 
  46 & Mean (SD) & 5.54 (2.14) \\ 
  47 & Median [Range] & 5.00 [3.00, 11.00] \\ 
  48 & \_\_\_Parity\_\_\_ &  \\ 
  49 & None & 111 / 600 (19\%) \\ 
  50 & One & 88 / 600 (15\%) \\ 
  51 & Two & 210 / 600 (35\%) \\ 
  52 & Three & 116 / 600 (19\%) \\ 
  53 & More than three & 75 / 600 (13\%) \\ 
  54 & \_\_\_Children\_\_\_ &  \\ 
  55 & Mean (SD) & 1.98 (1.41) \\ 
  56 & Median [Range] & 2.00 [0.00, 12.00] \\ 
  57 & \_\_\_Employment Status\_\_\_ &  \\ 
  58 & Unemployed & 104 / 600 (17\%) \\ 
  59 & Employed & 496 / 600 (83\%) \\ 
  60 & \_\_\_Educational Attainment\_\_\_ &  \\ 
  61 & High School or Less & 125 / 600 (21\%) \\ 
  62 & Some College & 190 / 600 (32\%) \\ 
  63 & College degree or higher & 285 / 600 (48\%) \\ 
  64 & \_\_\_Site\_\_\_ &  \\ 
  65 & B & 96 / 600 (16\%) \\ 
  66 & C & 76 / 600 (13\%) \\ 
  67 & M & 103 / 600 (17\%) \\ 
  68 & NJ & 27 / 600 (4.5\%) \\ 
  69 & P & 74 / 600 (12\%) \\ 
  70 & UCD & 102 / 600 (17\%) \\ 
  71 & UCLA & 122 / 600 (20\%) \\ 
   \hline
\end{tabular}
\end{table}
```

\newpage

## Table2: summary for AMH and visits

```{r}
#| label: "create_table2"
table2 <- swan_amh2 %>%
  ungroup() %>%
  dplyr::select(etime,
                fmp_age,
                amh,
                amh_min,
                amh_max,
                visit_n)  %>%
  tbl_summary(# by = site,
    missing = "ifany",
    label = list(c("fmp_age") ~ "Age of FMP (years)",
                 c("etime") ~ "Time of FMP (years)",
                 c("amh") ~ "Baseline AMH",
                 c("amh_min") ~ "Minimum AMH",
                 c("amh_max") ~ "Maximum AMH",
                 c("visit_n") ~ "Number of Visits"),
    type = list(all_continuous() ~ "continuous2",
                c("visit_n") ~ "continuous2"),
    statistic = list(all_continuous() ~ 
                       c("{mean} ({sd})",
                         # "{median} [{min}, {p5}, {p25}, {p75}, {p95}, {max}]",
                         "{median} [{p25}, {p75}]",
                         "{median} [{p5}, {p95}]",
                         "{median} [{min}, {max}]"))) %>%
  modify_header(label = "**Characteristics**") %>%
  # update the column header
  bold_labels() %>%
  # add_overall() %>%
  italicize_labels()
```

```{r}
#| label: "styled_table2"
#| tbl-cap: "Descriptive Statistics for SWAN AMH Data"

set_flextable_defaults(fonts_ignore = TRUE)

table2 %>% 
  as_flex_table() %>%
  fontsize(size = 8)
  
```

```{r "table2_latex"}
#| label: "table2_latex"
#| eval: false
library(xtable)
xtable(as.data.frame(table2), type = "latex",
      file = paste0("figure/swan_amh_09_publish_table2_", Sys.Date(), ".tex"))
```

Here is the latex code for Table2:

```         
% latex table generated in R 4.2.2 by xtable 1.8-4 package
% Wed Apr 10 21:00:42 2024
\begin{table}[ht]
\centering
\begin{tabular}{rll}
  \hline
 & **Characteristics** & **N = 600** \\ 
  \hline
  1 & \_\_\_Time of FMP (years)\_\_\_ &  \\ 
  2 & Mean (SD) & 10.98 (2.51) \\ 
  3 & Median [IQR] & 11.05 [9.33, 12.61] \\ 
  4 & Median [5\%, 95\%] & 11.05 [6.83, 15.10] \\ 
  5 & Median [Range] & 11.05 [3.59, 20.10] \\ 
  6 & \_\_\_Age of FMP (years)\_\_\_ &  \\ 
  7 & Mean (SD) & 52.98 (2.51) \\ 
  8 & Median [IQR] & 53.05 [51.33, 54.61] \\ 
  9 & Median [5\%, 95\%] & 53.05 [48.83, 57.10] \\ 
  10 & Median [Range] & 53.05 [45.59, 62.10] \\ 
  11 & \_\_\_Baseline AMH\_\_\_ &  \\ 
  12 & Mean (SD) & 540 (702) \\ 
  13 & Median [IQR] & 273 [71, 672] \\ 
  14 & Median [5\%, 95\%] & 273 [1, 2,095] \\ 
  15 & Median [Range] & 273 [1, 4,652] \\ 
  16 & \_\_\_Minimum AMH\_\_\_ &  \\ 
  17 & Mean (SD) & 28 (78) \\ 
  18 & Median [IQR] & 2 [1, 18] \\ 
  19 & Median [5\%, 95\%] & 2 [1, 133] \\ 
  20 & Median [Range] & 2 [1, 772] \\ 
  21 & \_\_\_Maximum AMH\_\_\_ &  \\ 
  22 & Mean (SD) & 628 (730) \\ 
  23 & Median [IQR] & 341 [132, 836] \\ 
  24 & Median [5\%, 95\%] & 341 [23, 2,174] \\ 
  25 & Median [Range] & 341 [1, 4,681] \\ 
  26 & \_\_\_Number of Visits\_\_\_ &  \\ 
  27 & Mean (SD) & 5.54 (2.14) \\ 
  28 & Median [IQR] & 5.00 [4.00, 7.00] \\ 
  29 & Median [5\%, 95\%] & 5.00 [3.00, 10.00] \\ 
  30 & Median [Range] & 5.00 [3.00, 11.00] \\ 
   \hline
\end{tabular}
\end{table}
```

-   [x] `r praise::praise()`
