==> 09_table1_publication.Rmd <==
---
title: "Table1 (for Publication)"
author: "randy"
date: "`r Sys.Date()`"
output:
  pdf_document:
    toc: true
    number_sections: true
  word_document:
    toc: yes
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE,
                      eval = TRUE,
                      warnings = FALSE,
                      message = FALSE,
                      comment = "#>",
                      cache = TRUE,
                      #results = "hide",
                      digits = 4,
                      error = FALSE)

## clean the R environment
# graphics.off()
# rm(list = ls())
# freshr::freshr()

## load packages
library(here, quietly = TRUE)
library(tidyverse, quietly = TRUE)
library(gtsummary, quietly = TRUE)
library(flextable, quietly = TRUE)
library(broom.mixed)
library(nlme)
library(JMbayes2)
devtools::load_all()

## check the directory for the file
# here::dr_here()
here::set_here()

## the figure or results should be saved 
# paste0("foldername/Sfilename_workingresult_", 
#      Sys.Date(), ".filetype")
```

\newpage

# To-do-list

-   [x] run table 1 for the total sample size as 1 group

    -   [x] Keeping table1 for one group (n = 600)

    -   [x] Baseline levels of AMH (untransformed)

    -   [x] also include the min and max of AMH

    -   [x] How many repeated measures per woman

-   [x] all the survival parameters summary

-   [x] all the longitudinal parameters summary

-   [x] joint model summary

    -   [x] with rate

    -   [x] without rate

\newpage

# Table1: descriptive summaries

Here is the Table1 for publication.

```{r "loading_data"}
load(here("data", "swan_amh1.rda"))
load(here("data", "swan_amh0.rda"))

sum_amh0 <- swan_amh0 %>%
  group_by(id) %>%
  summarize(visit_n = n(),
            amh_min = min(amh),
            amh_max = max(amh)) %>%
  ungroup() %>%
  mutate(id = as.character(id),
         visit_n = as.numeric(visit_n))

swan_amh2 <- swan_amh1 %>%
  full_join(sum_amh0, by = "id") %>%
  mutate(id = as.character(id),
         visit_n = as.numeric(visit_n))

```

```{r}
#| label: "create_table1"

table1 <- swan_amh2 %>%
  ungroup() %>%
  mutate(children = as.numeric(children),
         parity = case_when(parity == 0 ~ "None",
                            parity == 1 ~ "One",
                            parity == 2 ~ "Two",
                            parity == 3 ~ "Three",
                            parity == 4 ~ "More than three"),
         parity = factor(parity, levels = c("None", "One",
                                               "Two", "Three",
                                               "More than three"))) %>%
  dplyr::select(age, fmp_age, etime, 
                ethnic, married,
                alcohol, smoke, bc_pills,
                bmi, health, phys_act, amh,
                visit_n,  # visit,
                parity, children, 
                employed, edu, site)  %>%
  tbl_summary(# by = site,
              missing = "ifany",
              label = list(c("age") ~ "Age (years)",
                           c("fmp_age") ~ "Age of FMP (years)",
                           c("etime") ~ "Time of FMP(years)",
                           c("ethnic") ~ "Ethnicity / Race",
                           c("married") ~ "Marital Status",
                           c("alcohol") ~ "Alcohol Use",
                           c("smoke") ~ "Smoking History",
                           c("bc_pills") ~ "Ever use of BC Pills",
                           c("bmi") ~ "BMI",
                           c("health") ~ "Overall Health",
                           c("phys_act") ~ "Physical Activity Score",
                           c("amh") ~ "Baseline AMH",
                           c("parity") ~ "Parity",
                           c("children") ~ "Children",
                           c("employed") ~ "Employment Status",
                           c("edu") ~ "Educational Attainment",
                           c("visit_n") ~ "Number of Visits",
                           c("site") ~ "Site"),
              type = list(all_continuous() ~ "continuous2",
                          c("children", "visit_n") ~ "continuous2"),
              statistic = list(all_continuous() ~ c("{mean} ({sd})",
                                                    "{median} [{min}, {max}]"),
                               all_categorical() ~ "{n} / {N} ({p}%)")) %>%
  modify_header(label = "**Characteristics**") %>%
  # update the column header
  bold_labels() %>%
  # add_overall() %>%
  italicize_labels()
```

```{r}
#| label: "styled_table1"
#| tbl-cap: "Descriptive Statistics for SWAN AMH Data"
#| warning: false

set_flextable_defaults(fonts_ignore = TRUE)
# set_flextable_defaults(font.size = 8, 
#                        # text.align = "left",
#                        # theme_fun = theme_zebra,
#                        padding = 2)
table1 %>% 
  as_flex_table() %>%
  fontsize(size = 8)
  
```

Here is the LaTeX code for the table1.

```{r "table1_latex"}
#| label: "table_latex"

library(xtable)
xtable(as.data.frame(table1), type = "latex",
      file = paste0("figure/swan_amh_09_publish_table1_", Sys.Date(), ".tex"))
```


\newpage

# Table2: summary for AMH and visits

```{r}
#| label: "create_table2"
table2 <- swan_amh2 %>%
  ungroup() %>%
  dplyr::select(etime,
                fmp_age,
                amh,
                amh_min,
                amh_max,
                visit_n)  %>%
  tbl_summary(# by = site,
    missing = "ifany",
    label = list(c("fmp_age") ~ "Age of FMP (years)",
                 c("etime") ~ "Time of FMP (years)",
                 c("amh") ~ "Baseline AMH",
                 c("amh_min") ~ "Minimum AMH",
                 c("amh_max") ~ "Maximum AMH",
                 c("visit_n") ~ "Number of Visits"),
    type = list(all_continuous() ~ "continuous2",
                c("visit_n") ~ "continuous2"),
    statistic = list(all_continuous() ~ 
                       c("{mean} ({sd})",
                         # "{median} [{min}, {p5}, {p25}, {p75}, {p95}, {max}]",
                         "{median} [{p25}, {p75}]",
                         "{median} [{p5}, {p95}]",
                         "{median} [{min}, {max}]"))) %>%
  modify_header(label = "**Characteristics**") %>%
  # update the column header
  bold_labels() %>%
  # add_overall() %>%
  italicize_labels()
```

```{r}
#| label: "styled_table2"
#| tbl-cap: "Descriptive Statistics for SWAN AMH Data"

set_flextable_defaults(fonts_ignore = TRUE)

table2 %>% 
  as_flex_table() %>%
  fontsize(size = 8)
  
```

```{r "table2_latex"}
#| label: "table2_latex"

xtable(as.data.frame(table2), type = "latex",
      file = paste0("figure/swan_amh_09_publish_table2_", Sys.Date(), ".tex"))
```

Here is the latex code for Table2:


<!-- \newpage -->

<!-- # Longitudinal model -->

<!-- $lme.fit1: \ \ \  (fixed = lamh \sim time, \ random = \sim time| id)$ -->

<!-- Here is the summary for longitudinal sub-model, based on model selection -->

<!-- ```{r} -->
<!-- #| label: "dummy_variables" -->
<!-- swan_amh12 <- swan_amh11 %>% -->
<!--   cbind(model.matrix(~ 0 + site_ethnic + smoke,  -->
<!--                                     data = .)) %>% -->
<!--   janitor::clean_names() %>%  -->
<!--   dplyr::select(order(colnames(.))) -->

<!-- swan_amh02 <- swan_amh00 %>% -->
<!--   mutate(smoke = case_when(smoke == 1 ~ "Past Only", -->
<!--                            smoke == 2 ~ "Current Smoker", -->
<!--                            smoke == 3 ~ "Never Smoked"), -->
<!--          smoke = factor(smoke, levels = c("Never Smoked" , "Past Only", "Current Smoker"))) %>% -->
<!--   cbind(model.matrix(~ 0 + site_ethnic + smoke,  -->
<!--                                     data = .)) %>% -->
<!--   janitor::clean_names() %>%  -->
<!--   dplyr::select(order(colnames(.))) -->


<!-- control = lmeControl(opt = "optim") -->
<!-- lme.fit1 <- lme(lamh ~ time, -->
<!--                 data = swan_amh00, -->
<!--                 control = control, -->
<!--                 random = ~ time| id) -->
<!-- ``` -->

<!-- Here is the table for the longitudinal model: -->

<!-- ```{r} -->
<!-- broom.mixed::tidy(lme.fit1) %>% knitr::kable() -->
<!-- ``` -->

<!-- The full summary for the longitudinal model is as follows: -->

<!-- ```{r} -->
<!-- summary(lme.fit1) -->
<!-- ``` -->

<!-- \newpage -->

<!-- # Survival model -->

<!-- **Notes: is it necessary to include a model with observed AMH as covariate in survival model?** -->

<!-- $surv.fit16: \ \ \  coxph(Surv(etime, event) \sim married + bc\_pills + bmi + factor(site) + factor(ethnic) + factor(smoke))$ -->

<!-- Here is the table for survival model, based on model selection: -->

<!-- ```{r} -->
<!-- #| label: "survival_model" -->
<!-- surv.fit16 <- coxph(Surv(etime, event) ~ 0 + married + bc_pills + bmi + -->
<!--                       site_ethnic_b_black + # site_ethnic_b_cauca +  -->
<!--                       site_ethnic_c_black + site_ethnic_c_cauca +  -->
<!--                       site_ethnic_m_black + site_ethnic_m_cauca + -->
<!--                       site_ethnic_nj_cauca + site_ethnic_nj_hispa + -->
<!--                       site_ethnic_p_black + site_ethnic_p_cauca +  -->
<!--                       site_ethnic_ucd_cauca + site_ethnic_ucd_chine + -->
<!--                       site_ethnic_ucla_cauca + site_ethnic_ucla_japan + -->
<!--                       smoke_past_only + smoke_current_smoker, -->
<!--                     data = swan_amh12,  -->
<!--                     x = TRUE, -->
<!--                     model = TRUE) -->
<!-- ``` -->

<!-- ```{r} -->
<!-- broom::tidy(surv.fit16) %>% knitr::kable() -->

<!-- ``` -->

<!-- Here is the summary for the survival model -->

<!-- ```{r} -->
<!-- summary(surv.fit16) -->

<!-- ``` -->

<!-- \newpage -->

<!-- # Joint model without Rate (only value term of longitudinal submodel) -->

<!-- $$ -->
<!-- \begin{split} -->
<!--   jmbayes2.fit116: & \\  -->
<!--   lme.fit1: \ \ \  & \pmb m(time) = (fixed = lamh \sim time, \ random = \sim time| id) \\ -->
<!--   surv.fit16: \ \ \ & Surv(etime, event) \sim married + bc\_pills + bmi + factor(site) + \\ -->
<!--   & factor(ethnic) + factor(smoke) + \pmb m(time) -->
<!-- \end{split} -->
<!-- $$ -->

<!-- ## Longitudinal submodel -->

<!-- ```{r} -->
<!-- #| label: "joint_model_no_rate" -->
<!-- jmcontrol <- list(n_iter = 5000,  -->
<!--                   n_burnin = 1000,  -->
<!--                   n_chains = 1, -->
<!--                   seed = 55555, -->
<!--                   cores = 1) -->

<!-- jmbayes2.fit116 <- JMbayes2::jm(Surv_object = surv.fit16, -->
<!--                                Mixed_objects = lme.fit1,  -->
<!--                                time_var = "time", -->
<!--                                control = jmcontrol)  -->
<!-- ``` -->

<!-- Here is the summary for longitudinal submodel: -->

<!-- ```{r} -->
<!-- summary(jmbayes2.fit116)$Outcome1 %>% knitr::kable() -->
<!-- ``` -->

<!-- ## Survival submodel -->

<!-- Here is the summary for survival submodel: -->

<!-- ```{r} -->
<!-- summary(jmbayes2.fit116)$Survival %>% knitr::kable() -->
<!-- ``` -->

<!-- ## Joint model -->

<!-- Here are the full summary for the joint model without rate: -->

<!-- ```{r} -->
<!-- summary(jmbayes2.fit116) -->

<!-- ``` -->

<!-- \newpage -->

<!-- # Joint model with rate (value and rate terms of longitudinal submodel) -->

<!-- $$ -->
<!-- \begin{split} -->
<!--   jmbayes2.fit116.d: & \\ -->
<!--   lme.fit1: \ \ \  & \pmb m(time) = (fixed = lamh \sim time, \ random = \sim time| id) \\ -->
<!--   surv.fit16: \ \ \ & Surv(etime, event) \sim married + bc\_pills + bmi + \\ -->
<!--   & factor(site) + factor(ethnic) + factor(smoke) + \pmb m(time) + \pmb m'(time) -->
<!-- \end{split} -->
<!-- $$ -->

<!-- ```{r} -->
<!-- jmbayes2.fit116.d <- JMbayes2::jm(Surv_object = surv.fit16, -->
<!--                                   Mixed_objects = lme.fit1,  -->
<!--                                   functional_forms = list("lamh" = ~ value(lamh) + slope(lamh)), -->
<!--                                   time_var = "time", -->
<!--                                   control = jmcontrol, -->
<!--                                   seed = 55555)  -->

<!-- ``` -->

<!-- ## Longitudinal submodel -->

<!-- Here is the summary for longitudinal submodel: -->

<!-- ```{r} -->
<!-- summary(jmbayes2.fit116.d)$Outcome1 %>% knitr::kable() -->
<!-- ``` -->

<!-- ## Survival submodel -->

<!-- Here is the summary for survival submodel: -->

<!-- ```{r} -->
<!-- summary(jmbayes2.fit116.d)$Survival %>% knitr::kable() -->
<!-- ``` -->

<!-- ## Joint model -->

<!-- Here are the full summary for the joint model with rate: -->

<!-- ```{r} -->
<!-- summary(jmbayes2.fit116.d) -->
<!-- ``` -->

<!-- \newpage -->

<!-- # Comparison for model with and without rate -->

<!-- ```{r} -->
<!-- #| label: "compare_jmbayes2" -->

<!-- compare_jm(jmbayes2.fit116.d, jmbayes2.fit116) -->
<!-- ``` -->
