==> 15_auc_ci.qmd <==
---
title: "AUC CI"
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE,
                      warnings = FALSE,
                      message = FALSE,
                      comment = "#>",
                      cache = TRUE,
                      #results = "hide",
                      digits = 4,
                      error = FALSE)

# ## clean the R environment
# graphics.off()
# rm(list = ls())
# freshr::freshr()

## load packages
library(here, quietly = TRUE)
library(tidyverse, quietly = TRUE)
library(plyr, quietly = TRUE)
library(JMbayes2, quietly = TRUE)
library(auRoc, quietly = TRUE)
devtools::load_all()

## check the directory for the file
# here::dr_here()
here::set_here()

```

```{css, "css-setup", echo=FALSE}
.scroll-100 {
  max-height: 300px;
  max-width: 1000px;
  overflow-y: auto;
  background-color: inherit;
}
```

## To-do list

-   [x] Fit the joint model based on model selection with rate?

-   [x] confidence interval with the AUC

-   [x] `r praise::praise()`

## Value and Rate model

$$
\begin{split}
  jmb2.117.value.rate: & \\ 
  lme1: \ \ \  & \pmb m(time) = (fixed = lamh \sim time, \ random = \sim time| id) \\
  surv17: \ \ \ & Surv(etime, event) \sim phys.act + bmi + site + \\
  & ethnic + smoke + \pmb m(time) + \pmb m'(time) 
\end{split}
$$

```{r}
control = lmeControl(opt = "optim")
jmcontrol <- list(n_iter = 50000, 
                  n_burnin = 20000, 
                  n_chains = 1,
                  n_thin = 10,
                  seed = 55555,
                  cores = 1)

lme1 <- lme(lamh ~ time,
                data = swan_amh03,
                control = control,
                random = ~ time| id)

surv17 <- coxph(Surv(etime, event) ~ 
                  0 + # alcohol +
                  bmi + # bc_pills + 
                  # children +
                  # edu + 
                  # employed +
                  # health +
                  # married + 
                  phys_act + 
                  ethnic_black + ethnic_chine + ethnic_hispa + ethnic_japan + 
                  site_c + site_m + site_nj + site_p + site_ucd + site_ucla + 
                  smoke_current_smoker + smoke_past_only,
                data = swan_amh13, 
                x = TRUE,
                model = TRUE)

jmb2.117.value.rate <- JMbayes2::jm(Surv_object = surv17,
                                  Mixed_objects = lme1, 
                                  functional_forms = list("lamh" = ~ value(lamh) + slope(lamh)),
                                  time_var = "time",
                                  control = jmcontrol) 

summary117.value.rate <- summary(jmb2.117.value.rate)$Survival 

summary117.value.rate %>% knitr::kable(digits = 4)

```

## Function for AUC CI

This is function modified from Dimitris Rizopoulos' `tvAUC.jm()` function. The function is used to calculate the AUC and its confidence interval. The function is modified to work with the `jm` object from the `JMbayes2` package.

The function is conditional on the `Tstart` and `Dt` arguments for each setup. After we set up the `Tstart` and `Dt`, the new dataset will be generated to include the individuals who have not experienced the event yet but having longitudinal outcomes before the `Tstart` + `Dt`. We pull out every single iteration of `mcmc` object from the `predict()` function and calculate the AUC for that iteration.

argument `iter` is the number of iterations to calculate the AUC.

Then we can calculate the quantiles of the resampleing AUCs to get the confidence interval.

```{r}
tvAUC_CI <- function (object, 
                      newdata, 
                      Tstart,
                      Thoriz = NULL, 
                      Dt = NULL, 
                      iter = 1000L,
                      ...) {
  
  if (!inherits(object, "jm"))
    stop("Use only with 'jm' objects.\n")
  if (!is.data.frame(newdata) || nrow(newdata) == 0)
    stop("'newdata' must be a data.frame 
         with more than one rows.\n")
  if (is.null(Thoriz) && is.null(Dt))
    stop("either 'Thoriz' or 
         'Dt' must be non null.\n")
  if (!is.null(Thoriz) && Thoriz <= Tstart)
    stop("'Thoriz' must be larger than 'Tstart'.")
  if (is.null(Thoriz))
    Thoriz <- Tstart + Dt

  ## start of the dataset {{{---------------------------------------------------
  type_censoring <- object$model_info$type_censoring
  
  if (object$model_info$CR_MS)
    stop("'tvROC()' currently only works for right censored data.")
  
  Tstart <- Tstart + 1e-06
  Thoriz <- Thoriz + 1e-06
  id_var <- object$model_info$var_names$idVar
  time_var <- object$model_info$var_names$time_var
  Time_var <- object$model_info$var_names$Time_var
  event_var <- object$model_info$var_names$event_var
  
  if (is.null(newdata[[id_var]]))
    stop("cannot find the '", 
         id_var,
         "' variable in newdata.",
         sep = "")
  if (is.null(newdata[[time_var]]))
    stop("cannot find the '",
         time_var, 
         "' variable in newdata.",
         sep = "")
  if (any(sapply(Time_var, 
                 function (nmn) is.null(newdata[[nmn]]))))
    stop("cannot find the '", 
         paste(Time_var, 
               collapse = ", "),
         "' variable(s) in newdata.", 
         sep = "")
  if (is.null(newdata[[event_var]]))
    stop("cannot find the '", 
         event_var, 
         "' variable in newdata.", 
         sep = "")
  
  ## Need to find the newdata after the Tstart but 
  ## before the ind has any event yet
  newdata <- newdata[order(newdata[[Time_var]]), ]
  newdata <- newdata[newdata[[Time_var]] > Tstart, ]
  newdata <- newdata[newdata[[time_var]] <= Tstart, ]
  newdata[[id_var]] <- newdata[[id_var]][, drop = TRUE]
  test1 <- newdata[[Time_var]] < Thoriz & newdata[[event_var]] == 1
  
  if (!any(test1))
    stop("it seems that there are no events 
         in the interval [Tstart, Thoriz).")
  
  ## only predict for the people who does not 
  ## have any event yet but still has the 
  ## longitudinal outcomes before Tstart 
  newdata2 <- newdata
  newdata2[[Time_var]] <- Tstart
  newdata2[[event_var]] <- 0
  ## end of the dataset }}} ----------------------------------------------------
  
  
  ## prediction {{{-------------------------------------------------------------
  prediction.jm <- predict(object, 
                           newdata = newdata2, 
                           process = "event",
                           times = Thoriz, 
                           return_mcmc = TRUE,
                           n_samples = iter,
                           seed = 555L)
  
  data <- prediction.jm[["mcmc"]] %>% 
    as.data.frame()
  ## }}}------------------------------------------------------------------------
  
  preds_auc <- function(preds,
                        ...) {
    si_u_t <- 1 - preds
    names(si_u_t) <- prediction.jm$id
    si_u_t <- si_u_t[prediction.jm$times > Tstart]
    
    id <- newdata[[id_var]]
    Time <- newdata[[Time_var]]
    event <- newdata[[event_var]]
    f <- factor(id, levels = unique(id))
    
    ## time is the time for observation
    Time <- tapply(Time, f, tail, 1L)
    ## the event of the observation
    event <- tapply(event, f, tail, 1L)
    
    names(Time) <- names(event) <-  as.character(unique(id))
    
    if (any(dupl <- duplicated(Time))) {
      Time[dupl] <- Time[dupl] + 
        runif(length(Time[dupl]),
              1e-07,
              1e-06)}
    
    if (!all(names(si_u_t) == names(Time)))
      stop("mismatch between 'Time' variable names and
           survival probabilities names.")
    
    auc <- if (length(Time) > 1L) {
      ## why do you need pairs?
      pairs <- combn(as.character(unique(id)), 2)
      Ti <- Time[pairs[1, ]]
      Tj <- Time[pairs[2, ]]
      di <- event[pairs[1, ]]
      dj <- event[pairs[2, ]]
      si_u_t_i <- si_u_t[pairs[1, ]]
      si_u_t_j <- si_u_t[pairs[2, ]]
      
      
      ind1 <- (Ti <= Thoriz & di == 1) & Tj > Thoriz
      ind2 <- (Ti <= Thoriz & di == 0) & Tj > Thoriz
      ind3 <- (Ti <= Thoriz & di == 1) & (Tj <= Thoriz & dj == 0)
      ind4 <- (Ti <= Thoriz & di == 0) & (Tj <= Thoriz & dj == 0)
      
      names(ind1) <- 
        names(ind2) <- 
        names(ind3) <- 
        names(ind4) <-
        paste(names(Ti), 
              names(Tj),
              sep = "_")
      
      ind <- ind1 | ind2 | ind3 | ind4
      
      if (any(ind2)) {
        nams <- strsplit(names(ind2[ind2]), "_")
        nams_i <- sapply(nams, "[", 1)
        unq_nams_i <- unique(nams_i)
        preds2 <- predict(object, 
                          newdata = newdata[id %in% unq_nams_i, ],
                          process = "event", 
                          times = Thoriz, 
                          ...)
        pi_u_t <- preds2$pred
        f <- factor(preds2$id, levels = unique(preds2$id))
        names(pi_u_t) <- f
        pi_u_t <- tapply(pi_u_t, f, tail, 1)
        ind[ind2] <- ind[ind2] * pi_u_t[nams_i]
      }
      
      if (any(ind3)) {
        nams <- strsplit(names(ind3[ind3]), "_")
        nams_j <- sapply(nams, "[", 2)
        unq_nams_j <- unique(nams_j)
        preds3 <- predict(object,
                          newdata = newdata[id %in% unq_nams_j, ],
                          process = "event", 
                          times = Thoriz)
        qi_u_t <- preds3$pred
        f <- factor(preds3$id, 
                    levels = unique(preds3$id))
        names(qi_u_t) <- f
        qi_u_t <- 1 - tapply(qi_u_t,
                             f,
                             tail,
                             1)
        
        ind[ind3] <- ind[ind3] * qi_u_t[nams_j]
      }
      
      if (any(ind4)) {
        nams <- strsplit(names(ind4[ind4]), "_")
        nams_i <- sapply(nams, "[", 1)
        nams_j <- sapply(nams, "[", 2)
        unq_nams_i <- unique(nams_i)
        unq_nams_j <- unique(nams_j)
        preds4_i <- predict(object,
                            newdata = newdata[id %in% unq_nams_i, ],
                            process = "event", 
                            times = Thoriz, ...)
        pi_u_t <- preds4_i$pred
        f <- factor(preds4_i$id, 
                    levels = unique(preds4_i$id))
        names(pi_u_t) <- f
        pi_u_t <- tapply(pi_u_t, f, tail, 1)
        
        preds4_j <- predict(object, 
                            newdata = newdata[id %in% unq_nams_j, ],
                            process = "event", 
                            times = Thoriz, ...)
        qi_u_t <- preds4_j$pred
        f <- factor(preds4_j$id, 
                    levels = unique(preds4_j$id))
        names(qi_u_t) <- f
        qi_u_t <- 1 - tapply(qi_u_t, 
                             f, 
                             tail, 1)
        ind[ind4] <- ind[ind4] * pi_u_t[nams_i] * qi_u_t[nams_j]
      }
      sum((si_u_t_i < si_u_t_j) *
            c(ind), na.rm = TRUE) / 
        sum(ind, na.rm = TRUE)
    } else { NA }
    
    return(auc)
  }
  
  preds1 <- map(data, 
                ~preds_auc(preds = .x),
                .progress = TRUE)

  out <- list(auc = unlist(preds1),
              Tstart = Tstart, 
              Thoriz = Thoriz,
              # nr = length(unique(id)),
              classObject = class(object), 
              nameObject = deparse(substitute(object)))
  
  return(out)
}
```

```{r}
#| eval: false
AUC_s4_d2 <- tvAUC_CI(jmb2.117.value.rate,
                      newdata = swan_amh03,
                      Tstart = 4,
                      Dt = 2,
                      iter = 1000L)


AUC_s4_d2$auc %>% quantile(seq(0, 1, 0.05))

```

## Systematic AUC CI

```{r}
#| eval: false
#| include: true
AUC2  <- map(c(1, 2, 3, 4, 5), 
             ~try(tvAUC_CI(jmb2.117.value.rate,
                           swan_amh03,
                           Tstart = 2,
                           Dt = .x)))
AUC3  <- map(c(1, 2, 3, 4, 5), 
             ~try(tvAUC_CI(jmb2.117.value.rate,
                           swan_amh03,
                           Tstart = 3,
                           Dt = .x)))
AUC4  <- map(c(1, 2, 3, 4, 5), 
             ~try(tvAUC_CI(jmb2.117.value.rate,
                           swan_amh03,
                           Tstart = 4,
                           Dt = .x)))
AUC5  <- map(c(1, 2, 3, 4, 5), 
             ~try(tvAUC_CI(jmb2.117.value.rate,
                           swan_amh03,
                           Tstart = 5,
                           Dt = .x)))
AUC6  <- map(c(1, 2, 3, 4, 5), 
             ~try(tvAUC_CI(jmb2.117.value.rate,
                           swan_amh03,
                           Tstart = 6,
                           Dt = .x)))
AUC8  <- map(c(1, 2, 3, 4, 5), 
             ~try(tvAUC_CI(jmb2.117.value.rate,
                           swan_amh03,
                           Tstart = 8,
                           Dt = .x)))

save(AUC2, AUC3, AUC4, AUC5, AUC6, AUC8,
     file = "results/jmbayes2_119_value_rate_auc_ci.RData")
```

```{r}
load("results/jmbayes2_119_value_rate_auc_ci.RData")

AUC2_ci <- map_dfc(AUC2, "AUC") %>% View()
  map_dfr(~quantile(.x, c(0.5, 0.025, 0.975))) %>%
  round(digit = 3) %>% 
  unite(col = "95% CI", c("2.5%", "97.5%"), sep = " , ") %>%
  mutate(`95% CI` = paste0("(", `95% CI`, ")")) %>%
  unite(col = "AUC", c("50%", "95% CI"), sep = "  ") 

AUC3_ci <- map_dfc(AUC3, "auc") %>%
  map_dfr(~quantile(.x, c(0.5, 0.025, 0.975))) %>%
  round(digit = 3) %>%
  unite(col = "95% CI", c("2.5%", "97.5%"), sep = " , ") %>%
  mutate(`95% CI` = paste0("(", `95% CI`, ")")) %>%
  unite(col = "AUC", c("50%", "95% CI"), sep = "  ") 

AUC4_ci <- map_dfc(AUC4, "auc") %>%
  map_dfr(~quantile(.x, c(0.5, 0.025, 0.975))) %>%
  round(digit = 3) %>%
  unite(col = "95% CI", c("2.5%", "97.5%"), sep = " , ") %>%
  mutate(`95% CI` = paste0("(", `95% CI`, ")")) %>%
  unite(col = "AUC", c("50%", "95% CI"), sep = "  ") 

AUC5_ci <- map_dfc(AUC5, "auc") %>%
  map_dfr(~quantile(.x, c(0.5, 0.025, 0.975))) %>%
  round(digit = 3) %>%
  unite(col = "95% CI", c("2.5%", "97.5%"), sep = " , ") %>%
  mutate(`95% CI` = paste0("(", `95% CI`, ")")) %>%
  unite(col = "AUC", c("50%", "95% CI"), sep = "  ") 

AUC6_ci <- map_dfc(AUC6, "auc") %>%
  map_dfr(~quantile(.x, c(0.5, 0.025, 0.975))) %>%
  round(digit = 3) %>%
  unite(col = "95% CI", c("2.5%", "97.5%"), sep = " , ") %>%
  mutate(`95% CI` = paste0("(", `95% CI`, ")")) %>%
  unite(col = "AUC", c("50%", "95% CI"), sep = "  ") 

AUC8_ci <- map_dfc(AUC8, "auc") %>%
  map_dfr(~quantile(.x, c(0.5, 0.025, 0.975))) %>%
  round(digit = 3) %>%
  unite(col = "95% CI", c("2.5%", "97.5%"), sep = " , ") %>%
  mutate(`95% CI` = paste0("(", `95% CI`, ")")) %>%
  unite(col = "AUC", c("50%", "95% CI"), sep = "  ") 
```

```{r}
AUC2_ci <- rbind(NA, AUC2_ci)

AUC_final <- cbind(AUC2_ci, AUC3_ci,
                   AUC4_ci, AUC5_ci, 
                   AUC6_ci, AUC8_ci) 
colnames(AUC_final) <- c("Tstart2", "Tstart3", "Tstart4", "Tstart5", "Tstart6", "Tstart8")

AUC_final <- t(AUC_final)

colnames(AUC_final) <- c("Dt1", "Dt2", "Dt3", "Dt4", "Dt5")

```

For the final table each column is different `Dt` time window for predicting in $1, 2, 3, 4, 5$ years; each row is the `Tstart` time window for starting from the $2, 3, 4, 5, 6, 8$-th years. each cell is the AUC with 95% CI in the parenthesis `median (95% CI)`.

```{r}
AUC_final %>% 
  knitr::kable() %>%
  kableExtra::kable_styling(font_size = 8)

```

-   [x] `r praise::praise()`
