==> 26_2_permutation.qmd <==
---
title: "Resplite"
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE,
                      warnings = FALSE,
                      message = FALSE,
                      comment = "#>",
                      cache = TRUE,
                      #results = "hide",
                      digits = 4,
                      error = FALSE)

# ## clean the R environment
# graphics.off()
# rm(list = ls())
# freshr::freshr()

## load packages
library(here, quietly = TRUE)
library(tidyverse, quietly = TRUE)
library(JMbayes2, quietly = TRUE)

```

```{css, "css-setup", echo=FALSE}
.scroll-100 {
  max-height: 300px;
  max-width: 1000px;
  overflow-y: auto;
  background-color: inherit;
}

```

## To-do list

-   [x] re-split the data into training and testing 10 times

-   [x] fit the model and calculate the AUC

## Cross-validation

```{r}
## Install necessary packages if you don't have them
# install.packages("furrr")
# install.packages("future")  # If you don't have future package already

## Load the library
library(tidyverse)
library(JMbayes2)
# library(doParallel)
# library(foreach)

permutation <- function(seed = 555,
                        all_amh0 = all_amh0,
                        all_amh1 = all_amh1,
                        iteration,
                        burnin,
                        thin,
                        chain) {
  # browser()
  set.seed(seed)
  test_id <- sample(all_amh1$id, 300, 
                    replace = FALSE)
  test_amh1 <- all_amh1 %>% 
    filter(id %in% test_id)
  train_amh1 <- all_amh1 %>% 
    filter(!id %in% test_id)
  test_amh0 <- all_amh0 %>% 
    filter(id %in% test_id)
  train_amh0 <- all_amh0 %>% 
    filter(!id %in% test_id)
  
  control = lmeControl(opt = "optim")
  jmcontrol <- list(n_iter = iteration, 
                    n_burnin = burnin, 
                    n_thin = thin,
                    n_chains = chain,
                    save_random_effects = TRUE,
                    seed = seed,
                    cores = chain)
  lme1 <- lme(lamh ~ time,
              data = train_amh0,
              control = control,
              random = ~ time| id)
  
  surv19 <- coxph(Surv(etime, event) ~ 
                    0 + as.factor(bmi_cat)+ 
                    ethnic_black + ethnic_chine + ethnic_hispa + ethnic_japan + 
                    site_c + site_m + site_nj + site_p + site_ucd + site_ucla + 
                    smoke_current_smoker + smoke_past_only,
                  data = train_amh1, 
                  x = TRUE,
                  model = TRUE)
  
  jmb2.119.value.rate <- JMbayes2::jm(Surv_object = surv19,
                                      Mixed_objects = lme1, 
                                      functional_forms = list("lamh" = ~value(lamh) + slope(lamh)),
                                      time_var = "time",
                                      control = jmcontrol) 
  (total <- (iteration - burnin) / thin * chain)
  ## so we push the starting time into 45
  ## 45 is the 3rd year so the data should be similar to 
  ## the 3rd, 4th, and 5th year results plus one more
  AUC45  <- map(c(1, 2, 3, 4), 
               ~try(JMbayes2::tvAUC(jmb2.119.value.rate,
                                    train_amh0,
                                    Tstart = 2 + 3,
                                    seed = seed, Dt = .x)))
  AUC46  <- map(c(1, 2, 3, 4), 
               ~try(JMbayes2::tvAUC(jmb2.119.value.rate,
                                    train_amh0,
                                    Tstart = 3 + 3,
                                    seed = seed, Dt = .x)))
  AUC47  <- map(c(1, 2, 3, 4), 
               ~try(JMbayes2::tvAUC(jmb2.119.value.rate,
                                    train_amh0,
                                    Tstart = 4 + 3,
                                    seed = seed, Dt = .x)))
  AUC48  <- map(c(1, 2, 3, 4), 
               ~try(JMbayes2::tvAUC(jmb2.119.value.rate,
                                    train_amh0,
                                    Tstart = 5 + 3,
                                    seed = seed, Dt = .x)))
  AUC49  <- map(c(1, 2, 3, 4), 
               ~try(JMbayes2::tvAUC(jmb2.119.value.rate,
                                    train_amh0,
                                    Tstart = 6 + 3,
                                    seed = seed, Dt = .x)))
  # AUC8  <- map(c(1, 2, 3, 4), 
  #              ~try(JMbayes2::tvAUC(jmb2.119.value.rate,
  #                                   train_amh0,
  #                                   Tstart = 8,
  #                                   seed = seed, Dt = .x)))
  Tstart45 = map(AUC45, "auc") %>% as.character()
  Tstart46 = map(AUC46, "auc") %>% as.character()
  Tstart47 = map(AUC47, "auc") %>% as.character()
  Tstart48 = map(AUC48, "auc") %>% as.character()
  Tstart49 = map(AUC49, "auc") %>% as.character()
  # Tstart8 = map(AUC8, "auc") %>% as.character()
  tbl_auc_train <- rbind(Tstart45, 
                         Tstart46,
                         Tstart47,
                         Tstart48,
                         Tstart49) %>% 
    as.data.frame() %>%
    mutate_all(as.numeric) %>%
    round(4)
  names(tbl_auc_train) <- c("Dt1", "Dt2", "Dt3", "Dt4")
  
  AUC45_test  <- map(c(1, 2, 3, 4), 
                    ~try(JMbayes2::tvAUC(jmb2.119.value.rate,
                                         test_amh0,
                                         Tstart = 2 + 3,
                                         seed = seed, Dt = .x)))
  
  AUC46_test  <- map(c(1, 2, 3, 4), 
                    ~try(JMbayes2::tvAUC(jmb2.119.value.rate,
                                         test_amh0,
                                         Tstart = 3 + 3,
                                         seed = seed, Dt = .x)))
  AUC47_test  <- map(c(1, 2, 3, 4), 
                    ~try(JMbayes2::tvAUC(jmb2.119.value.rate,
                                         test_amh0,
                                         Tstart = 4 + 3,
                                         seed = seed, Dt = .x)))
  AUC48_test  <- map(c(1, 2, 3, 4), 
                    ~try(JMbayes2::tvAUC(jmb2.119.value.rate,
                                         test_amh0,
                                         Tstart = 5 + 3,
                                         seed = seed, Dt = .x)))
  AUC49_test  <- map(c(1, 2, 3, 4), 
                    ~try(JMbayes2::tvAUC(jmb2.119.value.rate,
                                         test_amh0,
                                         Tstart = 6 + 3,
                                         seed = seed, Dt = .x)))
  # AUC8_test  <- map(c(1, 2, 3, 4), 
  #                   ~try(JMbayes2::tvAUC(jmb2.119.value.rate,
  #                                        test_amh0,
  #                                        Tstart = 8,
  #                                        seed = seed, Dt = .x)))
  
  Tstart45_test = map(AUC45_test, "auc") %>% as.character()
  Tstart46_test = map(AUC46_test, "auc") %>% as.character()
  Tstart47_test = map(AUC47_test, "auc") %>% as.character()
  Tstart48_test = map(AUC48_test, "auc") %>% as.character()
  Tstart49_test = map(AUC49_test, "auc") %>% as.character()
  # Tstart8_test = map(AUC8_test, "auc") %>% as.character()
  tbl_auc_test <- rbind(Tstart45_test, 
                        Tstart46_test,
                        Tstart47_test,
                        Tstart48_test,
                        Tstart49_test) %>% 
    as.data.frame() %>%
    mutate_all(as.numeric) %>%
    round(4)
  names(tbl_auc_test) <- c("Dt1", "Dt2", "Dt3", "Dt4")

  
  Brier45  <- map(c(1, 2, 3, 4), 
                 ~try(JMbayes2::tvBrier(jmb2.119.value.rate,
                                        train_amh0,
                                        Tstart = 2 + 3,
                                        seed = seed, Dt = .x)))
  Brier46  <- map(c(1, 2, 3, 4), 
                 ~try(JMbayes2::tvBrier(jmb2.119.value.rate,
                                        train_amh0,
                                        Tstart = 3 + 3,
                                        seed = seed, Dt = .x)))
  Brier47  <- map(c(1, 2, 3, 4), 
                 ~try(JMbayes2::tvBrier(jmb2.119.value.rate,
                                        train_amh0,
                                        Tstart = 4 + 3,
                                        seed = seed, Dt = .x)))
  Brier48  <- map(c(1, 2, 3, 4), 
                 ~try(JMbayes2::tvBrier(jmb2.119.value.rate,
                                        train_amh0,
                                        Tstart = 5 + 3,
                                        seed = seed, Dt = .x)))
  Brier49  <- map(c(1, 2, 3, 4), 
                 ~try(JMbayes2::tvBrier(jmb2.119.value.rate,
                                        train_amh0,
                                        Tstart = 6 + 3,
                                        seed = seed, Dt = .x)))
  # Brier8  <- map(c(1, 2, 3, 4), 
  #                ~try(JMbayes2::tvBrier(jmb2.119.value.rate,
  #                                       train_amh0,
  #                                       Tstart = 8,
  #                                       seed = seed, Dt = .x)))
  
  Tstart45 = map(Brier45, "Brier") %>% as.character()
  Tstart46 = map(Brier46, "Brier") %>% as.character()
  Tstart47 = map(Brier47, "Brier") %>% as.character()
  Tstart48 = map(Brier48, "Brier") %>% as.character()
  Tstart49 = map(Brier49, "Brier") %>% as.character()
  # Tstart8 = map(Brier8, "Brier") %>% as.character()
  tbl_bs_train <- rbind(Tstart45,
                        Tstart46,
                        Tstart47,
                        Tstart48,
                        Tstart49) %>% 
    as.data.frame() %>%
    mutate_all(as.numeric) %>%
    round(4)
  names(tbl_bs_train) <- c("Dt1", "Dt2", "Dt3", "Dt4")
  
  Brier45_test  <- map(c(1, 2, 3, 4), 
                      ~try(JMbayes2::tvBrier(jmb2.119.value.rate,
                                             test_amh0,
                                             Tstart = 2 + 3,
                                             seed = seed, Dt = .x)))
  Brier46_test  <- map(c(1, 2, 3, 4), 
                      ~try(JMbayes2::tvBrier(jmb2.119.value.rate,
                                             test_amh0,
                                             Tstart = 3 + 3,
                                             seed = seed, Dt = .x)))
  Brier47_test  <- map(c(1, 2, 3, 4), 
                      ~try(JMbayes2::tvBrier(jmb2.119.value.rate,
                                             test_amh0,
                                             Tstart = 4 + 3,
                                             seed = seed, Dt = .x)))
  Brier48_test  <- map(c(1, 2, 3, 4), 
                      ~try(JMbayes2::tvBrier(jmb2.119.value.rate,
                                             test_amh0,
                                             Tstart = 5 + 3,
                                             seed = seed, Dt = .x)))
  Brier49_test  <- map(c(1, 2, 3, 4), 
                      ~try(JMbayes2::tvBrier(jmb2.119.value.rate,
                                             test_amh0,
                                             Tstart = 6 + 3,
                                             seed = seed, Dt = .x)))
  # # Brier8_test  <- map(c(1, 2, 3, 4), 
  #                     ~try(JMbayes2::tvBrier(jmb2.119.value.rate,
  #                                            test_amh0,
  #                                            Tstart = 8,
  #                                            Dt = .x)))
  
  Tstart45_test = map(Brier45_test, "Brier") %>% as.character()
  Tstart46_test = map(Brier46_test, "Brier") %>% as.character()
  Tstart47_test = map(Brier47_test, "Brier") %>% as.character()
  Tstart48_test = map(Brier48_test, "Brier") %>% as.character()
  Tstart49_test = map(Brier49_test, "Brier") %>% as.character()
  # Tstart8_test = map(Brier8_test, "Brier") %>% as.character()
  
  tbl_bs_test <- rbind(Tstart45_test,
                       Tstart46_test,
                       Tstart47_test,
                       Tstart48_test,
                       Tstart49_test) %>% 
    as.data.frame() %>%
    mutate_all(as.numeric) %>%
    round(4)
  names(tbl_bs_test) <- c("Dt1", "Dt2", "Dt3", "Dt4")
  
  return(amh_result = list(auc_train = tbl_auc_train,
                           auc_test = tbl_auc_test,
                           bs_train = tbl_bs_train,
                           bs_test = tbl_bs_test))
}

load("data/all_amh0.rda")
load("data/all_amh1.rda")
```

```{r}
library(foreach)
for (i in 46:100) {
  map_amh <- permutation(seed = i,
                         all_amh0 = all_amh0,
                         all_amh1 = all_amh1,
                         iteration = 150000,
                         burnin = 50000,
                         thin = 100,
                         chain = 5)
  save(map_amh, file = paste0("results/auc4/joint_model_seed_", i, ".Rdata"))
}
```
